Introduce SparkTool, add separate BuildBuckets tool. Register more classes for Kryo serialization. Represent coverage as Short instead of Int to save space. Make InnerRoutines an object.

+++ src/main/scala/dbpart/package.scala
@@ -7,7 +7,7 @@ package object dbpart {
    */
   type NTSeq = String
 
-  type Coverage = Int
+  type Coverage = Short
   
   /**
    * Type of edges between buckets in the macro graph.
